+++ src/main/java/com/beour/global/exception/GlobalExceptionHandler.java
@@ -5,7 +5,7 @@ import com.beour.global.exception.exceptionType.InputInvalidFormatException;
 import com.beour.global.exception.exceptionType.ReviewCommentNotFoundException;
 import com.beour.global.exception.exceptionType.UnauthorityException;
 import com.beour.global.response.ErrorResponse;
-import com.beour.reservation.commons.exceptionType.MissMatch;
+import com.beour.global.exception.exceptionType.MissMatch;
 import jakarta.validation.ConstraintViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
