+++ src/main/java/com/beour/global/exception/exceptionType/AvailableTimeNotFound.java
@@ -1,12 +1,12 @@
-package com.beour.reservation.commons.exceptionType;
+package com.beour.global.exception.exceptionType;
 
 import com.beour.global.exception.error.ErrorCode;
 
-public class MissMatch extends RuntimeException {
+public class AvailableTimeNotFound extends RuntimeException {
 
     private final Integer errorCode;
 
-    public MissMatch(ErrorCode error) {
+    public AvailableTimeNotFound(ErrorCode error) {
         super(error.getMessage());
         this.errorCode = error.getCode();
     }
@@ -14,5 +14,4 @@ public class MissMatch extends RuntimeException {
     public Integer getErrorCode(){
         return this.errorCode;
     }
-
 }
