+++ src/main/java/com/beour/global/exception/exceptionType/MissMatch.java
@@ -1,12 +1,12 @@
-package com.beour.reservation.commons.exceptionType;
+package com.beour.global.exception.exceptionType;
 
 import com.beour.global.exception.error.ErrorCode;
 
-public class ReservationNotFound extends RuntimeException{
+public class MissMatch extends RuntimeException {
 
     private final Integer errorCode;
 
-    public ReservationNotFound(ErrorCode error) {
+    public MissMatch(ErrorCode error) {
         super(error.getMessage());
         this.errorCode = error.getCode();
     }
