+++ src/main/java/com/beour/global/exception/exceptionType/ReservationNotFound.java
@@ -1,12 +1,12 @@
-package com.beour.reservation.commons.exceptionType;
+package com.beour.global.exception.exceptionType;
 
 import com.beour.global.exception.error.ErrorCode;
 
-public class AvailableTimeNotFound extends RuntimeException {
+public class ReservationNotFound extends RuntimeException{
 
     private final Integer errorCode;
 
-    public AvailableTimeNotFound(ErrorCode error) {
+    public ReservationNotFound(ErrorCode error) {
         super(error.getMessage());
         this.errorCode = error.getCode();
     }
@@ -14,4 +14,5 @@ public class AvailableTimeNotFound extends RuntimeException {
     public Integer getErrorCode(){
         return this.errorCode;
     }
+
 }
