+++ src/main/java/com/beour/reservation/host/service/ReservationCalendarService.java
@@ -6,12 +6,12 @@ import com.beour.global.exception.error.errorcode.UserErrorCode;
 import com.beour.global.exception.exceptionType.SpaceNotFoundException;
 import com.beour.global.exception.exceptionType.UnauthorityException;
 import com.beour.global.exception.exceptionType.UserNotFoundException;
-import com.beour.reservation.commons.exceptionType.MissMatch;
+import com.beour.global.exception.exceptionType.MissMatch;
 import com.beour.reservation.host.dto.CalendarReservationPageResponseDto;
 import com.beour.reservation.host.dto.CalendarReservationResponseDto;
 import com.beour.reservation.commons.entity.Reservation;
 import com.beour.reservation.commons.enums.ReservationStatus;
-import com.beour.reservation.commons.exceptionType.ReservationNotFound;
+import com.beour.global.exception.exceptionType.ReservationNotFound;
 import com.beour.reservation.commons.repository.ReservationRepository;
 import com.beour.space.domain.entity.Space;
 import com.beour.space.domain.repository.SpaceRepository;
