+++ src/main/java/com/beour/review/guest/service/ReviewGuestService.java
@@ -10,8 +10,8 @@ import com.beour.global.exception.exceptionType.UserNotFoundException;
 import com.beour.global.file.ImageUploader;
 import com.beour.reservation.commons.entity.Reservation;
 import com.beour.reservation.commons.enums.ReservationStatus;
-import com.beour.reservation.commons.exceptionType.MissMatch;
-import com.beour.reservation.commons.exceptionType.ReservationNotFound;
+import com.beour.global.exception.exceptionType.MissMatch;
+import com.beour.global.exception.exceptionType.ReservationNotFound;
 import com.beour.reservation.commons.repository.ReservationRepository;
 import com.beour.review.domain.entity.Review;
 import com.beour.review.domain.entity.ReviewComment;
@@ -37,8 +37,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
