+++ src/test/java/com/beour/reservation/guest/service/ReservationGuestServiceTest.java
@@ -7,9 +7,9 @@ import com.beour.global.exception.exceptionType.SpaceNotFoundException;
 import com.beour.reservation.commons.entity.Reservation;
 import com.beour.reservation.commons.enums.ReservationStatus;
 import com.beour.reservation.commons.enums.UsagePurpose;
-import com.beour.reservation.commons.exceptionType.AvailableTimeNotFound;
-import com.beour.reservation.commons.exceptionType.MissMatch;
-import com.beour.reservation.commons.exceptionType.ReservationNotFound;
+import com.beour.global.exception.exceptionType.AvailableTimeNotFound;
+import com.beour.global.exception.exceptionType.MissMatch;
+import com.beour.global.exception.exceptionType.ReservationNotFound;
 import com.beour.reservation.commons.repository.ReservationRepository;
 import com.beour.reservation.guest.dto.ReservationCreateRequest;
 import com.beour.reservation.guest.dto.ReservationListPageResponseDto;
