+++ src/test/java/com/beour/reservation/host/service/ReservationCalendarServiceTest.java
@@ -8,8 +8,8 @@ import com.beour.global.exception.exceptionType.UserNotFoundException;
 import com.beour.reservation.commons.entity.Reservation;
 import com.beour.reservation.commons.enums.ReservationStatus;
 import com.beour.reservation.commons.enums.UsagePurpose;
-import com.beour.reservation.commons.exceptionType.MissMatch;
-import com.beour.reservation.commons.exceptionType.ReservationNotFound;
+import com.beour.global.exception.exceptionType.MissMatch;
+import com.beour.global.exception.exceptionType.ReservationNotFound;
 import com.beour.reservation.commons.repository.ReservationRepository;
 import com.beour.reservation.host.dto.CalendarReservationPageResponseDto;
 import com.beour.reservation.host.dto.CalendarReservationResponseDto;
